Extract movement into seperate file
